Builds SessionFactory using DI
