+++ runtime/src/test/java/com/flipkart/flux/dao/StateMachinesDAOTest.java
@@ -13,7 +13,7 @@
 
 package com.flipkart.flux.dao;
 
-import com.flipkart.flux.HibernateModule;
+import com.flipkart.flux.guice.HibernateModule;
 import com.flipkart.flux.dao.iface.StateMachinesDAO;
 import com.flipkart.flux.domain.State;
 import com.flipkart.flux.domain.StateMachine;
